Revert "Runnable locals(#9007)"
This reverts commit 1c64db575c16bc3a3ec5e815bf5f39d385238222.
