+++ libs/langchain/langchain/schema/runnable/__init__.py
@@ -1,4 +1,3 @@
-from langchain.schema.runnable._locals import GetLocalVar, PutLocalVar
 from langchain.schema.runnable.base import (
     Runnable,
     RunnableBinding,
@@ -12,8 +11,6 @@ from langchain.schema.runnable.passthrough import RunnablePassthrough
 from langchain.schema.runnable.router import RouterInput, RouterRunnable
 
 __all__ = [
-    "GetLocalVar",
-    "PutLocalVar",
     "RouterInput",
     "RouterRunnable",
     "Runnable",
