+++ libs/langchain/tests/unit_tests/schema/test_runnable.py
@@ -132,24 +132,15 @@ async def test_default_method_implementations(mocker: MockerFixture) -> None:
         ["hello", "wooorld"], [dict(tags=["a-tag"]), dict(metadata={"key": "value"})]
     ) == [5, 7]
     assert spy.call_args_list == [
-        mocker.call(
-            "hello", dict(tags=["a-tag"], metadata={}, callbacks=None, _locals={})
-        ),
-        mocker.call(
-            "wooorld",
-            dict(metadata={"key": "value"}, tags=[], callbacks=None, _locals={}),
-        ),
+        mocker.call("hello", dict(tags=["a-tag"])),
+        mocker.call("wooorld", dict(metadata={"key": "value"})),
     ]
     spy.reset_mock()
 
     assert fake.batch(["hello", "wooorld"], dict(tags=["a-tag"])) == [5, 7]
     assert spy.call_args_list == [
-        mocker.call(
-            "hello", dict(tags=["a-tag"], metadata={}, callbacks=None, _locals={})
-        ),
-        mocker.call(
-            "wooorld", dict(tags=["a-tag"], metadata={}, callbacks=None, _locals={})
-        ),
+        mocker.call("hello", dict(tags=["a-tag"])),
+        mocker.call("wooorld", dict(tags=["a-tag"])),
     ]
     spy.reset_mock()
 
@@ -170,14 +161,8 @@ async def test_default_method_implementations(mocker: MockerFixture) -> None:
         7,
     ]
     assert spy.call_args_list == [
-        mocker.call(
-            "hello",
-            dict(metadata={"key": "value"}, tags=[], callbacks=None, _locals={}),
-        ),
-        mocker.call(
-            "wooorld",
-            dict(metadata={"key": "value"}, tags=[], callbacks=None, _locals={}),
-        ),
+        mocker.call("hello", dict(metadata={"key": "value"})),
+        mocker.call("wooorld", dict(metadata={"key": "value"})),
     ]
